projeto-comparador/requirements.txt
python-dotenv==1.0.1
PyGithub==2.5.0
groq==0.9.0
